votesy-web/requirements.txt
Flask==3.0.2
gunicorn==20.1.0
flask-requests=0.0.14
pika==1.3.2